feat: remove deprecated svelteBracketNewLine
BREAKING CHANGE: a previous commit introduced an enforced rule to set
the order of blocks in Svelte files. The same order was used in all our
components since we started with Svelte 4y ago. We need to keep it.
However, some projects may alreadt have adopted the new order. So to not
break them, we need to branch the version.